+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/RabbitMq/RabbitMqLegacyReceiveTests.cs
@@ -10,7 +10,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.RabbitMq
 {
     [NetFrameworkTest]
     public class RabbitMqLegacyReceiveTests : IClassFixture<RemoteServiceFixtures.RabbitMqLegacyReceiverFixture>
